medium workable state attack.js and winner.js are created
